+++ Changelog.md
@@ -11,6 +11,7 @@
 - Adds support for different font render modes: `lcd`, `light`, `gray`, `monochrome` in `profiles.NAME.font.render_mode` (default: `lcd`).
 - Adds VT sequence for querying/setting current font `OSC 50 ; ? ST` and `OSC 50 ; Font ST`.
 - Adds experimental text reflow.
+- Adds OpenFileManager action to configuration.
 - ...
 
 ### 0.1.1 (2020-12-31)